Removes hard dependency of `riptide-micrometer` to `riptide-failsafe` in Spring auto configuration
Fixes gh-774
