+++ riptide-spring-boot-autoconfigure/src/main/java/org/zalando/riptide/autoconfigure/MicrometerFailsafeFactory.java
@@ -1,0 +1,39 @@
+package org.zalando.riptide.autoconfigure;
+
+import com.google.common.collect.ImmutableList;
+import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Tag;
+import org.zalando.riptide.failsafe.CircuitBreakerListener;
+import org.zalando.riptide.failsafe.CompositeRetryListener;
+import org.zalando.riptide.failsafe.LoggingRetryListener;
+import org.zalando.riptide.failsafe.RetryListener;
+import org.zalando.riptide.failsafe.metrics.MetricsCircuitBreakerListener;
+import org.zalando.riptide.failsafe.metrics.MetricsRetryListener;
+
+final class MicrometerFailsafeFactory {
+
+    private MicrometerFailsafeFactory() {
+
+    }
+
+    public static CircuitBreakerListener createCircuitBreakerListener(final MeterRegistry registry,
+            final ImmutableList<Tag> defaultTags) {
+        return new MetricsCircuitBreakerListener(registry).withDefaultTags(defaultTags);
+    }
+
+    public static CircuitBreakerListener getDefaultCircuitBreakerListener() {
+        return CircuitBreakerListener.DEFAULT;
+    }
+
+    public static RetryListener createRetryListener(final MeterRegistry registry,
+            final ImmutableList<Tag> defaultTags) {
+        return new CompositeRetryListener(
+                new MetricsRetryListener(registry).withDefaultTags(defaultTags),
+                new LoggingRetryListener()
+        );
+    }
+
+    public static RetryListener getDefaultRetryListener() {
+        return new LoggingRetryListener();
+    }
+}
