+++ riptide-spring-boot-autoconfigure/src/main/java/org/zalando/riptide/autoconfigure/MicrometerPluginFactory.java
@@ -4,12 +4,6 @@ import com.google.common.collect.ImmutableList;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
 import org.zalando.riptide.Plugin;
-import org.zalando.riptide.failsafe.CircuitBreakerListener;
-import org.zalando.riptide.failsafe.CompositeRetryListener;
-import org.zalando.riptide.failsafe.LoggingRetryListener;
-import org.zalando.riptide.failsafe.RetryListener;
-import org.zalando.riptide.failsafe.metrics.MetricsCircuitBreakerListener;
-import org.zalando.riptide.failsafe.metrics.MetricsRetryListener;
 import org.zalando.riptide.micrometer.MicrometerPlugin;
 
 final class MicrometerPluginFactory {
@@ -21,26 +15,5 @@ final class MicrometerPluginFactory {
     public static Plugin createMicrometerPlugin(final MeterRegistry registry,
             final ImmutableList<Tag> tags) {
         return new MicrometerPlugin(registry).withDefaultTags(tags);
-    }
-
-    public static CircuitBreakerListener createCircuitBreakerListener(final MeterRegistry registry,
-            final ImmutableList<Tag> defaultTags) {
-        return new MetricsCircuitBreakerListener(registry).withDefaultTags(defaultTags);
-    }
-
-    public static CircuitBreakerListener getDefaultCircuitBreakerListener() {
-        return CircuitBreakerListener.DEFAULT;
-    }
-
-    public static RetryListener createRetryListener(final MeterRegistry registry,
-            final ImmutableList<Tag> defaultTags) {
-        return new CompositeRetryListener(
-                new MetricsRetryListener(registry).withDefaultTags(defaultTags),
-                new LoggingRetryListener()
-        );
-    }
-
-    public static RetryListener getDefaultRetryListener() {
-        return new LoggingRetryListener();
     }
 }
