+++ riptide-spring-boot-autoconfigure/src/test/java/org/zalando/riptide/autoconfigure/NoRiptideFailsafeTest.java
@@ -1,0 +1,16 @@
+package org.zalando.riptide.autoconfigure;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.util.ReflectionUtils;
+
+import java.lang.reflect.Method;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+final class NoRiptideFailsafeTest {
+    @Test
+    void shouldReadClassDefinitionWithoutFailsafePlugin() {
+        Method[] methods = ReflectionUtils.getUniqueDeclaredMethods(MicrometerPluginFactory.class);
+        assertNotNull(methods);
+    }
+}
